[REFACTOR] messagePublisher 중복 제거
