+++ src/main/java/com/express/application/service/user/UserService.java
@@ -9,9 +9,9 @@ import com.express.application.port.input.user.request.JoinUserRequest;
 import com.express.application.port.input.user.request.ModifyUserRequest;
 import com.express.application.port.input.user.response.ReadUserResponse;
 import com.express.application.port.output.inmemory.redis.RedisProcessor;
+import com.express.application.port.output.messaging.MessagePublisher;
 import com.express.application.port.output.user.UserProcessor;
 import com.express.application.port.output.user.UserReader;
-import com.express.application.service.messaging.MessagePublisher;
 import com.express.infrasturcture.common.UseCase;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
